Create AddToDoFragment and recreate same funcionality in AddToDoFragment from AddToDoActivity
