+++ app/src/main/res/layout/activity_add_to_do.xml
@@ -1,174 +1,11 @@
 <LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:animateLayoutChanges="true"
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:layout_width="match_parent"
     >
-    <include layout="@layout/base_toolbar"/>
-
-    <android.support.design.widget.CoordinatorLayout
-        android:layout_height="match_parent"
-        android:layout_width="match_parent"
-        >
-
-        <LinearLayout
-            android:orientation="vertical"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <LinearLayout
-                android:orientation="vertical"
-                android:background="?attr/colorPrimary"
-                android:id="@+id/editTextParentLinearLayout"
-                android:gravity="center"
-                android:layout_weight=".5"
-                android:layout_width="match_parent"
-                android:layout_height="0dp">
+    <include layout="@layout/layout_fragment_container"/>
 
 
-                <com.example.avjindersinghsekhon.minimaltodo.Utility.CustomTextInputLayout
-                    app:hintTextAppearance="@style/FloatingLabelTextAppearance"
-                    android:id="@+id/toDoCustomTextInput"
-                    android:gravity="center"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    >
-                    <EditText
-                        android:id="@+id/userToDoEditText"
-                        android:hint="@string/title"
-                        android:lines="2"
-                        android:inputType="textAutoCorrect|textCapSentences"
-                        android:maxLines="3"
-                        android:textColor="@color/icons"
-                        android:gravity="start|bottom"
-                        android:layout_marginRight="32dp"
-                        android:layout_marginLeft="32dp"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content" />
-
-                </com.example.avjindersinghsekhon.minimaltodo.Utility.CustomTextInputLayout>
-
-            </LinearLayout>
-            <LinearLayout
-                android:id="@+id/todoReminderAndDateContainerLayout"
-                android:layout_weight="1.5"
-                android:layout_width="match_parent"
-                android:orientation="vertical"
-                android:layout_height="0dp">
-
-                <LinearLayout
-                    android:orientation="horizontal"
-                    android:layout_weight=".5"
-                    android:gravity="center"
-                    android:layout_marginTop="32dp"
-                    android:layout_marginLeft="42dp"
-                    android:layout_marginRight="42dp"
-                    android:layout_width="match_parent"
-                    android:layout_height="0dp">
-
-                    <ImageButton
-                        android:layout_weight=".4"
-                        android:background="@android:color/transparent"
-                        android:id="@+id/userToDoReminderIconImageButton"
-                        android:padding="8dp"
-                        android:src="@drawable/ic_add_alarm_grey_200_24dp"
-                        android:layout_width="0dp"
-                        android:layout_height="wrap_content" />
-
-                    <TextView
-                        android:gravity="start"
-                        android:paddingLeft="8dp"
-                        android:paddingRight="8dp"
-                        android:paddingTop="8dp"
-                        android:paddingBottom="8dp"
-                        android:id="@+id/userToDoRemindMeTextView"
-                        android:text="@string/remind_me"
-                        android:textSize="16sp"
-                        android:textColor="@color/secondary_text"
-                        android:layout_weight="2"
-                        android:layout_width="0dp"
-                        android:layout_height="wrap_content" />
-
-                    <android.support.v7.widget.SwitchCompat
-                        android:id="@+id/toDoHasDateSwitchCompat"
-                        android:layout_gravity="center"
-                        android:padding="8dp"
-                        android:gravity="center"
-                        android:layout_weight="1"
-                        android:layout_width="0dp"
-                        android:layout_height="wrap_content" />
-
-                </LinearLayout>
-
-                <LinearLayout
-                    android:id="@+id/toDoEnterDateLinearLayout"
-                    android:animateLayoutChanges="true"
-                    android:layout_weight="1"
-                    android:layout_width="match_parent"
-                    android:layout_marginLeft="48dp"
-                    android:layout_marginRight="48dp"
-                    android:gravity="center"
-                    android:orientation="vertical"
-                    android:layout_height="0dp">
-
-                    <LinearLayout
-
-                        android:layout_width="match_parent"
-                        android:gravity="top"
-                        android:layout_weight="1"
-                        android:layout_height="0dp">
-
-                        <Button
-                            android:id="@+id/newToDoChooseDateButton"
-                            android:text="@string/choose_date"
-                            android:background="@drawable/button_pressed_background"
-                            android:textColor="@android:color/white"
-                            android:layout_weight="1"
-                            android:layout_marginRight="4dp"
-                            android:layout_width="0dp"
-                            android:layout_height="wrap_content" />
-
-
-                        <Button
-                            android:layout_marginLeft="4dp"
-                            android:id="@+id/newToDoChooseTimeButton"
-                            android:textColor="@android:color/white"
-                            android:text="@string/choose_time"
-                            android:background="@drawable/button_pressed_background"
-                            android:layout_weight="1"
-                            android:layout_width="0dp"
-                            android:layout_height="wrap_content" />
-                    </LinearLayout>
-
-                    <TextView
-                        android:layout_weight="2"
-                        android:id="@+id/newToDoDateTimeReminderTextView"
-                        android:textSize="14sp"
-                        android:gravity="start"
-                        android:text="@string/remind_date_and_time"
-                        android:textColor="@color/secondary_text"
-                        android:layout_width="match_parent"
-                        android:layout_height="0dp" />
-
-                </LinearLayout>
-
-
-            </LinearLayout>
-
-        </LinearLayout>
-
-        <android.support.design.widget.FloatingActionButton
-            android:id="@+id/makeToDoFloatingActionButton"
-            android:layout_marginEnd="16dp"
-            android:layout_marginRight="16dp"
-            android:src="@drawable/ic_send_white_18dp"
-            android:layout_height="wrap_content"
-            app:layout_anchor="@id/editTextParentLinearLayout"
-            app:layout_anchorGravity="bottom|right"
-            app:fabSize="normal"
-            android:layout_width="wrap_content"
-            />
-    </android.support.design.widget.CoordinatorLayout>
 </LinearLayout>
